perf(rl): batch off-policy CUDA workloads

+++ quant/wavelets.py
@@ -15,9 +15,9 @@ def rolling_wavelet_features(
     """Compute each row from its trailing window only."""
     if window < 2**level:
         raise ValueError("wavelet window is too short for the requested level")
-    values = series.astype(float).to_numpy(copy=True)
+    values = series.to_numpy(dtype=np.float32, copy=True)
     columns = [f"wavelet_a{level}_endpoint", *(f"wavelet_d{i}_energy" for i in range(level, 0, -1))]
-    output = np.full((len(values), len(columns)), np.nan, dtype=float)
+    output = np.full((len(values), len(columns)), np.nan, dtype=np.float32)
     if len(values) >= window:
         windows = np.lib.stride_tricks.sliding_window_view(values, window)
         coefficients = pywt.wavedec(windows, wavelet, level=level, mode="periodization", axis=-1)
